+++ src/Compute/Compute/Manual/VirtualMachineScaleSetCreateOrUpdateMethod.cs
@@ -306,6 +306,20 @@ namespace Microsoft.Azure.Commands.Compute.Automation
         [PSArgumentCompleter("Any", "Auto")]
         public string ZonePlacementPolicy { get; set; }
 
+        [Parameter(
+            ParameterSetName = SimpleParameterSet,
+            Mandatory = false,
+            HelpMessage = "This property supplements the 'zonePlacementPolicy' property. If 'zonePlacementPolicy' is set to 'Any', availability zone selected by the system must be present in the list of availability zones passed with 'includeZones'. If 'includeZones' is not provided, all availability zones in region will be considered for selection.")]
+        [ValidateNotNullOrEmpty]
+        public string[] IncludeZone { get; set; }
+
+        [Parameter(
+            ParameterSetName = SimpleParameterSet,
+            Mandatory = false,
+            HelpMessage = "This property supplements the 'zonePlacementPolicy' property. If 'zonePlacementPolicy' is set to 'Any', availability zone selected by the system must not be present in the list of availability zones passed with 'excludeZones'. If 'excludeZones' is not provided, all availability zones in region will be considered for selection.")]
+        [ValidateNotNullOrEmpty]
+        public string[] ExcludeZone { get; set; }
+
         private void ConfigureSecuritySettings()
         {
             if (SecurityType?.ToLower() == SecurityTypes.TrustedLaunch ||
@@ -568,7 +582,9 @@ namespace Microsoft.Azure.Commands.Compute.Automation
                     securityPostureExcludeExtension: _cmdlet.SecurityPostureExcludeExtension,
                     enableProxyAgent: _cmdlet.EnableProxyAgent ? true : (bool?)null,
                     addProxyAgentExtension: _cmdlet.AddProxyAgentExtension.IsPresent ? true : (bool?)null,
-                    zonePlacementPolicy: _cmdlet.ZonePlacementPolicy
+                    zonePlacementPolicy: _cmdlet.ZonePlacementPolicy,
+                    includeZone: _cmdlet.IncludeZone,
+                    excludeZone: _cmdlet.ExcludeZone
                     );
             }
 
@@ -710,7 +726,9 @@ namespace Microsoft.Azure.Commands.Compute.Automation
                     securityPostureExcludeExtension: _cmdlet.SecurityPostureExcludeExtension,
                     enableProxyAgent: _cmdlet.EnableProxyAgent ? true : (bool?)null,
                     addProxyAgentExtension: _cmdlet.AddProxyAgentExtension.IsPresent ? true : (bool?)null,
-                    zonePlacementPolicy: _cmdlet.ZonePlacementPolicy
+                    zonePlacementPolicy: _cmdlet.ZonePlacementPolicy,
+                    includeZone: _cmdlet.IncludeZone,
+                    excludeZone: _cmdlet.ExcludeZone
                 );
             }
         }